# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/me:
    get:
      operationId: getMe
      tags: [profiles]
      summary: Get current user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: updateMe
      tags: [profiles]
      summary: Update current user profile (e.g. password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List profiles (scoped to caller's tenant and org unit)
      responses:
        "200":
          description: Profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateProfile
      tags: [profiles]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete profile
      responses:
        "204":
          description: Deleted
  /profiles/{id}/passkeys:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listProfilePasskeys
      tags: [profiles]
      summary: List passkeys for a profile
      responses:
        "200":
          description: Passkeys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PasskeyCredential"
  /profiles/{id}/passkeys/{credentialId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: credentialId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteProfilePasskey
      tags: [profiles]
      summary: Delete a passkey
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Right:
      type: string
      enum:
        - manage_profiles
        - assign_roles
        - manage_roles
        - manage_org_units
        - create_checklists
        - edit_checklists
        - delete_checklists
        - view_reports
        - manage_checklist_settings
    OrgUnitType:
      type: string
      enum: [region, district, location]
    OrgUnit:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        parentId:
          type: integer
          nullable: true
        name:
          type: string
        type:
          $ref: "#/components/schemas/OrgUnitType"
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, name, type, createdAt]
...
    Role:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
          nullable: true
        name:
          type: string
        level:
          type: integer
        isSystem:
          type: boolean
        rights:
          type: array
          items:
            $ref: "#/components/schemas/Right"
      required: [id, name, level, isSystem, rights]
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
          nullable: true
        orgUnitId:
          type: integer
          nullable: true
        orgUnit:
          $ref: "#/components/schemas/OrgUnit"
          nullable: true
        username:
          type: string
        displayName:
          type: string
        roleId:
          type: integer
        role:
          $ref: "#/components/schemas/Role"
        mustChangePassword:
          type: boolean
        isActive:
          type: boolean
      required: [id, username, displayName, roleId, role, mustChangePassword, isActive]
    CreateProfileRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        displayName:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 12
          pattern: '^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$'
        roleId:
          type: integer
        orgUnitId:
          type: integer
          nullable: true
        mustChangePassword:
          type: boolean
      required: [username, displayName, password, roleId]
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        displayName:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 12
          pattern: '^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$'
        roleId:
          type: integer
        orgUnitId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
    UpdateMeRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
          pattern: '^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$'
...
    PasskeyCredential:
      type: object
      properties:
        id:
          type: integer
        credentialId:
          type: string
        label:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, credentialId, createdAt]
...