# lib/api-zod/src/generated/types/createProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProfileRequest {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  displayName: string;
  /**
   * @minLength 12
   * @pattern ^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$
   */
  password: string;
  roleId: number;
  orgUnitId?: number | null;
  mustChangePassword?: boolean;
}